requirements.txt
flask
gunicorn
google-cloud-bigquery
google-cloud-core
